Refactor mina proof task service
