+++ packages/mina-service/src/domain/index.ts
@@ -1,2 +1,2 @@
-export * from './create-proof';
+export * from './proof';
 export * from './zk-compile';
