+++ packages/mina-service/src/service/task-service.ts
@@ -1,11 +1,12 @@
 import { logger } from '@helper';
-import { createProof } from '@domain';
+import { Proof } from '@domain';
+import { withCompoundTransaction } from '@zkdb/storage';
 
 export class TaskService {
   private maxRetries: number;
   private initialDelay: number;
 
-  constructor(maxRetries: number = 5, initialDelay: number = 1000) {
+  constructor(initialDelay: number = 1000, maxRetries: number = Infinity) {
     this.maxRetries = maxRetries;
     this.initialDelay = initialDelay;
   }
@@ -14,43 +15,48 @@ export class TaskService {
     return new Promise((resolve) => setTimeout(resolve, ms));
   }
 
-  public async fetchAndProcessTasks(): Promise<void> {
+  public async run(): Promise<void> {
     let retries = 0;
     let delay = this.initialDelay;
 
     while (true) {
-      // TODO: Perform get next task. We don't call broker anymore
-      const result = await getNextTaskId();
+      if (retries >= this.maxRetries) {
+        break;
+      }
 
-      if (result.type === 'success') {
-        const id = result.data;
-
-        if (id) {
-          logger.info('Task received with id:', id);
-
-          try {
-            await createProof(id);
-          } catch (error) {
-            logger.error(error);
-          }
-
-          retries = 0; // Reset retries on success
-        } else {
-          logger.info('No task available, waiting...');
-          await this.delay(delay);
-          delay = Math.min(delay * 2, 32000); // Exponential backoff with cap
-          delay += Math.floor(Math.random() * 1000); // Add jitter
-          retries++;
+      const noTaskFound = await withCompoundTransaction(async (session) => {
+        const result = await Proof.getNextTask(session);
+        logger.info('Task received with id:', result);
+        if (result === null) {
+          return true;
         }
-      } else {
-        logger.error('Error processing task:', result.message);
+
+        try {
+          await Proof.create(result, session);
+        } catch (error) {
+          logger.error(`Error processing task with ID ${result._id}: ${error}`);
+        }
+
+        return false;
+      });
+
+      if (noTaskFound) {
+        logger.info('No task available, waiting...');
         await this.delay(delay);
         delay = Math.min(delay * 2, 32000); // Exponential backoff with cap
         delay += Math.floor(Math.random() * 1000); // Add jitter
-        retries += 1;
+        retries++;
+      } else {
+        delay = this.initialDelay;
+        retries = 0;
       }
     }
 
-    logger.info('Max retries reached, stopping task fetching.');
+    logger.info('Maximum retries reached. Stopping task consumption.');
   }
 }
+
+export const TASK_SERVICE = {
+  clusterName: 'task',
+  payload: new TaskService().run,
+};
